Don't create VOLUME for /opt/codescene
The purpose of a volume is to persist data between docker
images. There is no reason to do this for the CodeScene executable,
and in fact this creates problems when trying to debug. It seems that
ADD, which normally diffs the existing file and the new file to decide
if it should be updated or not, doesn't have the same effect when
copying to  a volume. The result was that even when rebuilding an image,
CodeScene was not getting updated, which made debugging much more
difficult.