Upgrade to Ubuntu 18.10 and OpenJDK 11 for better memory utilization (-XX:MaxRAMPercentage=60)
In Java 10, better support for containers has been introduced which
includes `-XX:MaxRAMPercentage` - we set this to 60%.
This is useful because otherwise Java uses only 1/4 of total heap in the
container which is way too low given that running codescene is the solo
purpose of this container.
Moreover, in Java 10+ the `UseContainerSupport` is on by default.
In Java 8 we would have to use experimental option
`-XX:+UseCGroupMemoryLimitForHeap` which still allows us to use only
`XX:MaxRAMFraction` option

Note that `-XX:MaxRAMPercentage` has no effect is user specified `-Xmx`
explicitly.

We also install full JDK (not just JRE) which includes useful
debugging/monitoring tools like `jcmd`. It does, however, make the image
bigger (+200 MB).